pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tracelens"
version = "1.0.0"
description = "Enhanced traceroute with automatic network intelligence enrichment"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "TraceLens Team"}
]
keywords = ["traceroute", "network", "asn", "geolocation", "diagnostics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: System :: Networking",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",
    "dnspython>=2.4.0",
    "httpx>=0.25.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/tracelens"
Repository = "https://github.com/yourusername/tracelens"
Issues = "https://github.com/yourusername/tracelens/issues"

[project.scripts]
tracelens = "tracelens.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tracelens*"]

[tool.setuptools.package-data]
tracelens = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
